BackEnd/requirements.txt
django>=5,<6
djangorestframework
djangorestframework-simplejwt
django-tenants
psycopg2-binary
gunicorn
django-cors-headers
drf-spectacular
celery
redis
django-redis
django-environ
channels
channels-redis
requests
fedapay
sendgrid-django
python-dotenv
geonamescache
reportlab
matplotlib
PyJWT>=2,<3
python-docx
openpyxl
python-pptx
